requirements.txt
asttokens==3.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
clearml==2.0.2
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
furl==2.1.4
idna==3.11
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.2.6
orderedmultidict==1.0.2
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathlib2==2.3.7.post1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
